CI-335 part way through working out the page flows for search

+++ src/Web/Sfa.Das.Sas.Core/Domain/Model/ProviderSearchResultSummary.cs
@@ -8,7 +8,7 @@ namespace Sfa.Das.Sas.Core.Domain.Model
 {
     public class ProviderSearchResultSummary
     {
-        public int UkPrn { get; set; }
+        public long UkPrn { get; set; }
         public string ProviderName { get; set; }
     }
 }
